requirements.txt
flask==3.0.2
pandas==2.2.1
python-pptx
python-docx
Pillow
lxml>=4.9.0 
